remove dead code, fix invalid imports, and final cleanup before merge
