+++ app/api/feed/types.ts
@@ -1,5 +1,19 @@
 export type FeedItemType = 'blog' | 'video' | 'farcaster'
 
+export type FarcasterCastData = {
+  hash: string
+  author?: { fid?: number; username?: string; display_name?: string; pfp_url?: string }
+  text?: string
+  timestamp?: string
+  embeds?: {
+    images?: { url: string }[]
+    videos?: { thumbnailUrl?: string; imageUrl?: string }[]
+    urls?: { openGraph?: { image?: string } }[]
+  }
+  reactions?: { likes_count: number; recasts_count: number }
+  replies?: { count: number }
+}
+
 export type FeedItem = {
   id: string
   type: FeedItemType
@@ -10,5 +24,5 @@ export type FeedItem = {
   text?: string // For Farcaster posts without images
   url: string
   description?: string
-  castData?: any // Raw Farcaster cast data for FarcasterEmbed component
+  castData?: FarcasterCastData
 }
