+++ package.json
@@ -13,8 +13,7 @@
   },
   "dependencies": {
     "@farcaster/miniapp-sdk": "^0.2.1",
-    "@google/model-viewer": "^4.1.0",
-    "@react-three/drei": "^10.7.7",
+"@react-three/drei": "^10.7.7",
     "@react-three/fiber": "^9.4.2",
     "@tailwindcss/postcss": "4.1.17",
     "@tanstack/react-query": "^5.90.10",
