--- types/model-viewer.d.ts
@@ -1,20 +1,0 @@
-import 'react'
-
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      'model-viewer': React.DetailedHTMLProps<
-        React.HTMLAttributes<HTMLElement> & {
-          src: string
-          'auto-rotate'?: boolean
-          'camera-controls'?: boolean
-          'interaction-policy'?: string
-        },
-        HTMLElement
-      >
-    }
-  }
-}
-
-export {}
-
